hot fix userid cast error

+++ Source-Codes/IERSystem/IERSystem/Areas/QuanLySoChuyenMau/Controllers/APIController.cs
@@ -24,7 +24,7 @@ namespace IERSystem.Areas.QuanLySoChuyenMau.Controllers
             {
                 try
                 {
-                    cacsonhanmau_inp.CreatedBy = db.Users.Find((int)Session["loggedId"]);
+                    cacsonhanmau_inp.CreatedBy = db.Users.Find((long)Session["loggedId"]);
                     CacSoChuyenMauAPIImpl.CreateModel(cacsonhanmau_inp, db);
                     db.SaveChanges();
                     Console.WriteLine("OK");
@@ -93,7 +93,7 @@ namespace IERSystem.Areas.QuanLySoChuyenMau.Controllers
             {
                 try
                 {
-                    mauptadd_inp.CreatedBy = db.Users.Find((int)Session["loggedId"]);
+                    mauptadd_inp.CreatedBy = db.Users.Find((long)Session["loggedId"]);
                     SoChuyenMauAPIImpl.AddMauPT(mauptadd_inp, db);
                     db.SaveChanges();
                     Console.WriteLine("OK");
