+++ Source-Codes/IERSystem/IERSystem/Areas/QuanLySoChuyenMau/Controllers/CacSoChuyenMauController.cs
@@ -28,7 +28,7 @@ namespace IERSystem.Areas.QuanLySoChuyenMau.Controllers
             if (ModelState.IsValid)
             {
                 var today = DateTime.Now;
-                cacsochuyenmau_inp.CreatedBy = db.Users.Find((int)Session["loggedID"]);
+                cacsochuyenmau_inp.CreatedBy = db.Users.Find((long)Session["loggedID"]);
                 cacsochuyenmau_inp.CreatedDate = DateTime.Now;
                 cacsochuyenmau_inp.Nam = today.Year;
                 cacsochuyenmau_inp.TuThang = today.Month;
